+++ index.html
@@ -1,6 +1,5 @@
 
 <html lang=en>
-
     <head>
         <meta charset=utf-8>
         <title>memeRoom</title>
